+++ apps/cow-fi/pages/404.tsx
@@ -1,35 +1,69 @@
 import Head from 'next/head'
-import Link from 'next/link'
 import { GetStaticProps } from 'next'
-import { CONFIG, SiteConfig } from '@/const/meta'
-import { Title, Section } from '@/components/Layout/index.styles'
+import { Color } from '@cowprotocol/ui'
+import styled from 'styled-components'
+import { CONFIG } from '@/const/meta'
 import Layout from '@/components/Layout'
 
-// pages/404.js
-export default function Custom404({ siteConfigData }: { siteConfigData: SiteConfig }) {
-  const { title } = siteConfigData
+import { ContainerCard, ArticleContent, ArticleMainTitle, BodyContent } from '@/styles/styled'
+
+const DATA_CACHE_TIME_SECONDS = 5 * 60 // Cache 5min
+
+interface PageProps {
+  siteConfigData: typeof CONFIG
+}
+
+const Wrapper = styled.div`
+  display: flex;
+  flex-flow: column wrap;
+  justify-content: flex-start;
+  align-items: center;
+  max-width: 1000px;
+  width: 100%;
+  margin: 24px auto 0;
+  gap: 24px;
+`
+
+export default function Page({ siteConfigData }: PageProps) {
+  const { title, descriptionShort } = siteConfigData
 
   return (
-    <>
+    <Layout bgColor={Color.neutral90}>
       <Head>
-        <title>Page Not Found (404) - {title}</title>
+        <title>{title}</title>
       </Head>
-      <Layout>
-        <Section>
-          <Title>404 - Page Not Found</Title>
-          <p>
-            This page could not be found. Please go back to the <Link href="/">home page.</Link>
-          </p>
-        </Section>
-      </Layout>
-    </>
+
+      <Wrapper>
+        <ContainerCard bgColor={'transparent'} minHeight="70vh" gap={62} gapMobile={42} centerContent touchFooter>
+          <ArticleContent maxWidth="100%">
+            <ArticleMainTitle margin={'0 0 62px'} fontSize={52}>
+              {title}
+            </ArticleMainTitle>
+
+            <BodyContent>
+              <p>
+                This page could not be found. Please go back to the <a href="/">homepage</a> or use the navigation menu
+                to find what you are looking for.
+              </p>
+            </BodyContent>
+          </ArticleContent>
+        </ContainerCard>
+      </Wrapper>
+    </Layout>
   )
 }
 
-export const getStaticProps: GetStaticProps = async () => {
+export const getStaticProps: GetStaticProps<PageProps> = async () => {
   const siteConfigData = CONFIG
 
   return {
-    props: { siteConfigData },
+    props: {
+      siteConfigData: {
+        ...siteConfigData,
+        title: '404 - Page Not Found',
+        descriptionShort: '404 - Page Not Found',
+      },
+    },
+    revalidate: DATA_CACHE_TIME_SECONDS,
   }
 }
