+++ apps/cow-fi/pages/careers/index.tsx
@@ -8,7 +8,7 @@ import styled from 'styled-components'
 
 import { CONFIG } from '@/const/meta'
 
-import LayoutV2 from '@/components/Layout/LayoutV2'
+import Layout from '@/components/Layout'
 
 import {
   ContainerCard,
@@ -55,7 +55,7 @@ export default function Page({ siteConfigData, jobsData }: PageProps) {
   const jobsCountForDepartment = department === 'All' ? jobsCount : jobsData[department].length
 
   return (
-    <LayoutV2 bgColor={Color.neutral90}>
+    <Layout bgColor={Color.neutral90}>
       <Head>
         <title>
           {siteConfigData.title} - {siteConfigData.descriptionShort}
@@ -200,7 +200,7 @@ export default function Page({ siteConfigData, jobsData }: PageProps) {
           </ContainerCardSection>
         </ContainerCard>
       </Wrapper>
-    </LayoutV2>
+    </Layout>
   )
 }
