+++ apps/cow-fi/pages/tokens/[tokenId].tsx
@@ -1,12 +1,27 @@
 import React from 'react'
 import Head from 'next/head'
-import Layout from '@/components/Layout'
 import { getTokensIds as getTokensIds, getTokenDetails as getTokenDetails } from 'services/tokens'
 import { TokenDetails as TokenDetailsPure, TokenDetailProps } from '@/components/TokenDetails'
 import { GetStaticPaths, GetStaticProps } from 'next'
 import { CONFIG } from '@/const/meta'
 
+import Layout from '@/components/Layout'
+import { Color } from '@cowprotocol/ui'
+import styled from 'styled-components'
+
 const DATA_CACHE_TIME_SECONDS = 10 * 60 // 10 minutes
+
+const Wrapper = styled.div`
+  display: flex;
+  flex-flow: column wrap;
+  justify-content: flex-start;
+  align-items: center;
+  max-width: 1640px;
+  width: 100%;
+  margin: 56px auto;
+  gap: 24px;
+  font-size: 1.6rem;
+`
 
 export type TokenDetailPageProps = TokenDetailProps
 
@@ -29,8 +44,10 @@ export default function TokenDetailsPage({ token }: TokenDetailPageProps) {
         <meta key="twitterTitle" name="twitter:title" content={CONFIG.title} />
       </Head>
 
-      <Layout fullWidthGradientVariant={true}>
-        <TokenDetailsPure token={token} />
+      <Layout bgColor={Color.neutral90}>
+        <Wrapper>
+          <TokenDetailsPure token={token} />
+        </Wrapper>
       </Layout>
     </>
   )
